+++ hydra-cardano-api/src/Hydra/Cardano/Api/PlutusScript.hs
@@ -6,9 +6,7 @@ import Hydra.Cardano.Api.Prelude
 
 import qualified Cardano.Ledger.Alonzo.Language as Ledger
 import qualified Cardano.Ledger.Alonzo.Scripts as Ledger
-import Codec.Serialise (serialise)
-import Data.ByteString.Short (toShort)
-import qualified PlutusLedgerApi.V2 as Plutus
+import qualified PlutusLedgerApi.Common as Plutus
 
 -- * Type Conversions
 
@@ -18,7 +16,7 @@ import qualified PlutusLedgerApi.V2 as Plutus
 --
 -- (a) If the given script is a timelock script, it throws an impure exception;
 -- (b) If the given script is in a wrong language, it silently coerces it.
-fromLedgerScript :: HasCallStack => Ledger.Script era -> PlutusScript lang
+fromLedgerScript :: HasCallStack => Ledger.AlonzoScript era -> PlutusScript lang
 fromLedgerScript = \case
   Ledger.TimelockScript{} -> error "fromLedgerScript: TimelockScript"
   Ledger.PlutusScript _ bytes -> PlutusScriptSerialised bytes
@@ -28,17 +26,17 @@ toLedgerScript ::
   forall lang.
   (IsPlutusScriptLanguage lang) =>
   PlutusScript lang ->
-  Ledger.Script (ShelleyLedgerEra Era)
+  Ledger.AlonzoScript (ShelleyLedgerEra Era)
 toLedgerScript (PlutusScriptSerialised bytes) =
   let lang = case plutusScriptVersion @lang of
         PlutusScriptV1 -> Ledger.PlutusV1
         PlutusScriptV2 -> Ledger.PlutusV2
    in Ledger.PlutusScript lang bytes
 
--- | Convert a plutus 'Script' into a cardano-api 'PlutusScript'
-fromPlutusScript :: Plutus.Script -> PlutusScript lang
+-- | Convert a serialized plutus script into a cardano-api 'PlutusScript'.
+fromPlutusScript :: Plutus.SerialisedScript -> PlutusScript lang
 fromPlutusScript =
-  PlutusScriptSerialised . toShort . toStrict . serialise
+  PlutusScriptSerialised
 
 -- * Orphans